Reorganize: Move WelcomeController
Move WelcomeController to a Controllers package for organization purposes
