+++ src/main/java/com/battleship/backend/controllers/WelcomeController.java
@@ -1,4 +1,4 @@
-package com.battleship.backend;
+package com.battleship.backend.controllers;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
