+++ src/test/java/com/battleship/backend/BattleshipApplicationTests.java
@@ -1,5 +1,6 @@
 package com.battleship.backend;
 
+import com.battleship.backend.controllers.WelcomeController;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
